add constraints for tracking into cerebellum
Adding a mechanism to prevent streamlines directly jumping from brain stem to cerebellar gray matter. The present method may require some modifications depending on how the cerebellar meshes are constructed but should not be too difficult to make extensions...

+++ src/dwi/tractography/MACT/scenemodeller.h
@@ -58,7 +58,7 @@ class SceneModeller
     // methods to find the nearest tissue from a given point
     bool nearestTissue( const Eigen::Vector3d& point,
                         struct Intersection& intersection,
-                        size_t stride = std::numeric_limits< size_t >::max() ) const;
+                        int32_t stride = std::numeric_limits< int32_t >::max() ) const;
 
     // methods to check whether a point is inside/on a target tissue type
     bool inTissue( const Eigen::Vector3d& point, const TissueType& type,
